backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
hyperliquid-python-sdk
python-dotenv